Split WebAuthn request parsing support
